requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
openpyxl==3.1.2
langchain>=0.1.0
langchain-community>=0.0.24
langchain-groq>=0.1.1
chromadb==0.4.22
python-dotenv==1.0.0
numpy==1.26.0
sentence-transformers>=2.2.2
matplotlib>=3.8.0
statsmodels>=0.14.0
scikit-learn>=1.4.0